// src/api/model/appConfigResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Athelix_API
 * Backend API for workout tracking, exercise logging, personal record detection, mesocycle planning, and training analytics.
 * OpenAPI spec version: 0.1.0
 */
import type { AppAuthConfigResponse } from './appAuthConfigResponse';
import type { AppDocsConfigResponse } from './appDocsConfigResponse';
import type { AppFeaturesResponse } from './appFeaturesResponse';
import type { AppSupportedValuesResponse } from './appSupportedValuesResponse';

export interface AppConfigResponse {
  app_name: string;
  version: string;
  environment: string;
  auth: AppAuthConfigResponse;
  docs: AppDocsConfigResponse;
  features: AppFeaturesResponse;
  supported_values: AppSupportedValuesResponse;
}