// src/api/model/userResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Athelix_API
 * Backend API for workout tracking, exercise logging, personal record detection, mesocycle planning, and training analytics.
 * OpenAPI spec version: 0.1.0
 */

export interface UserResponse {
  id: number;
  username: string;
  /**
     * @minLength 6
     * @maxLength 254
     * @pattern ^[A-Za-z0-9.!#$%&'*+/=?^_`{|}~-]+@[A-Za-z0-9-]+(?:\.[A-Za-z0-9-]+)+$
     */
  email: string;
  created_at: string;
  updated_at: string;
}